Layout complete.
The layout is done.  I am happy with the scalability of all the sections now.  It could be better still in places but it is better than I thought I was capable of just a few days ago

Next is tidying the code and giving credit to the photographers in the readme.